requirements.txt
alembic==1.4.3
aniso8601==8.1.0
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
croniter==0.3.36
decorator==4.4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Humanize==0.3.0
Flask-JWT==0.3.2
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
flask-rest-paginate==1.1.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-RQ2==18.3
Flask-SQLAlchemy==2.4.4
humanize==3.1.0
infinity==1.5
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
natsort==7.1.0
passlib==1.7.4
phonenumbers==8.12.14
Pillow==8.0.1
pycparser==2.20
PyJWT==1.4.2
PyMySQL==0.10.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
redis==3.5.3
rq==1.7.0
rq-scheduler==0.10.0
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
validators==0.18.1
Werkzeug==0.16.0
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
WTForms-JSON==0.3.3
